backend/requirements.txt
# Core
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ClipsAI & Dependencies
clipsai
whisperx @ git+https://github.com/m-bain/whisperx.git
ffmpeg-python==0.2.0

# Optional (if adding OpenAI later)
openai==1.12.0